Added new Sale sign used to sell regions.
Permissions: regionsigns.create.sale and regionsigns.use.sale

The maker of the sign must own the region the sign is for, or they need the regionsigns.use.norestrict perm.

When the region is purchased, the proceeds are distributed evenly to the players listed as owners for the region. After that, the region's owners and members are cleared and the new owner set.